# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de gestion de Gestion Promos\nPOUR UNE FACILITATION DES TESTS, L'AUTHENTIFICATION EST TEMPORAIREMENT DÉSACTIVÉE !",
        "version": "1.0.0",
        "description": "API Gestion-Promos"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": ["authentification"],
                "summary": "Authentification de l'utilisateur",
                "description": "Endpoint pour authentifier un utilisateur avec un email et un mot de passe.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "required": ["email", "password"]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Authentification réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Requête invalide",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email ou mot de passe manquant"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentification échouée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Email ou mot de passe incorrect"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "get": {
                "tags": ["users"],
                "summary": "Lister tous les utilisateurs",
                "description": "Recupére la liste de tous les utilisateurs",
                "responses": {
                    "200": {
                        "description": "Liste des utilisateurs",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": ["users"],
                "summary": "Créer un nouvel utilisateur",
                "description": "Endpoint pour créer un nouvel utilisateur avec les données spécifiées.",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "prenom": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "adresse": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "telephone": {
                                        "type": "string",
                                        "pattern": "^((77|76|75|70|78)\\d{3}\\d{2}\\d{2})|(33[8]\\d{2}\\d{2}\\d{2})$"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "Admin",
                                            "CM",
                                            "Manager",
                                            "Coach",
                                            "Apprenant"
                                        ]
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "maxLength": 255
                                    },
                                    "password": {
                                        "type": "string",
                                        "minLength": 6
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "Photo de profil à télécharger",
                                        "maxLength": 1024
                                    },
                                    "statut": {
                                        "type": "string",
                                        "enum": ["actif", "bloqué"]
                                    }
                                },
                                "required": [
                                    "nom",
                                    "prenom",
                                    "adresse",
                                    "telephone",
                                    "role",
                                    "email",
                                    "password"
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Utilisateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "nom": {
                                            "type": "string"
                                        },
                                        "prenom": {
                                            "type": "string"
                                        },
                                        "adresse": {
                                            "type": "string"
                                        },
                                        "telephone": {
                                            "type": "string"
                                        },
                                        "role": {
                                            "type": "string"
                                        },
                                        "email": {
                                            "type": "string"
                                        },
                                        "photo": {
                                            "type": "string",
                                            "format": "binary"
                                        },
                                        "statut": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/referentiels": {
            "get": {
                "tags": ["referentiels"],
                "summary": "Lister tous les référentiels actifs",
                "description": "Recupére la liste des référentiels",
                "responses": {
                    "200": {
                        "description": "Liste des référentiels",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Non autorisé",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            },
            "post": {
                "tags": ["referentiels"],
                "summary": "Créer un nouveau référentiel",
                "description": "Endpoint pour créer un nouveau référentiel.",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "codeReferentiel": {
                                        "type": "string",
                                        "description": "Code unique du référentiel"
                                    },
                                    "libelleReferentiel": {
                                        "type": "string",
                                        "description": "Libellé unique du référentiel"
                                    },
                                    "StatutReferentiel": {
                                        "type": "string",
                                        "description": "Statut du référentiel",
                                        "enum": ["Actif", "Inactif", "Archiver"]
                                    },
                                    "photo": {
                                        "type": "string",
                                        "format": "binary",
                                        "description": "Image du référentiel"
                                    },
                                    "competences": {
                                        "type": "array",
                                        "items": {
                                            "type": "object",
                                            "properties": {
                                                "nom": {
                                                    "type": "string",
                                                    "description": "Nom de la compétence"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "description": "Type de compétence",
                                                    "enum": [
                                                        "Back-end",
                                                        "Front-end"
                                                    ]
                                                }
                                            },
                                            "required": ["nom", "type"]
                                        }
                                    }
                                },
                                "required": [
                                    "codeReferentiel",
                                    "libelleReferentiel",
                                    "StatutReferentiel"
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Référentiel créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "codeReferentiel": {
                                            "type": "string"
                                        },
                                        "libelleReferentiel": {
                                            "type": "string"
                                        },
                                        "StatutReferentiel": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/referentiels/archive/referentiel": {
            "get": {
                "tags": ["referentiels"],
                "summary": "Lister les référentiels supprimés (archivés)",
                "description": "Ce endpoint permet de lister tous les référentiels qui ont été supprimés (ceux marqués avec le champ 'deleted_at').",
                "parameters": [
                    {
                        "name": "statut",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": ["archived"]
                        },
                        "description": "Optionnel: Filtrer les référentiels par statut, ex: 'archived' pour les archivés"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des référentiels archivés récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "codeReferentiel": {
                                                "type": "string"
                                            },
                                            "libelleReferentiel": {
                                                "type": "string"
                                            },
                                            "StatutReferentiel": {
                                                "type": "string"
                                            },
                                            "deleted_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Aucun référentiel archivé trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Archive Référentiel vide."
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}